Exchange: add special handling of XYZT coordinates
It is pretty clear that handling all these various attributes will be a rather
tedious task. For now there is a bunch of hard coding and repeating, these
should be factored out into functions whenever that these functions are
Become apparent.

The YAY news is, as of this change, these files are openable in ODV with
little need to change the default options.

It is also becoming a little more clear how to handle the cases where multiple
WHP names map to one cf name/netcdf variable. Using string array attributes
for this seems to be an “ok” option for now, might want to do a CFey “space
separated” text attribute in the future.